benchbuild & demo & dist updates
