--- alignmentbiascheckingtools/README.md
@@ -1,1 +1,0 @@
-<img width="1298" alt="abct" src="https://github.com/user-attachments/assets/b9e906c2-f6be-44b3-8276-6fc2d0bd1288">
